backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
pillow>=10.0
